+++ src/components/CustomResources/CustomResourcesByGroup.js
@@ -1,7 +1,6 @@
-import React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { Link } from 'shared/components/Link/Link';
-import { Link as RRLink } from 'react-router-dom';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
+import { Link } from 'react-router-dom';
 import pluralize from 'pluralize';
 import { GroupingListPage } from './GroupingListPage';
 import { useUrl } from 'hooks/useUrl';
@@ -11,14 +10,11 @@ export default function CustomResourcesByGroup({ enableColumnLayout }) {
   const { namespace, clusterUrl, scopedUrl } = useUrl();
   const description = (
     <Trans i18nKey="custom-resources.description">
-      <Link
+      <ExternalLink
         className="bsl-link"
         url="https://kubernetes.io/docs/concepts/extend-kubernetes/api-extension/custom-resources/"
       />
-      <RRLink
-        className="bsl-link"
-        to={clusterUrl(`customresourcedefinitions`)}
-      />
+      <Link className="bsl-link" to={clusterUrl(`customresourcedefinitions`)} />
     </Trans>
   );
