+++ src/resources/ConfigMaps/ConfigMapList.js
@@ -1,9 +1,8 @@
-import React from 'react';
 import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ControlledBy } from 'shared/components/ControlledBy/ControlledBy';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
 
 import { ConfigMapCreate } from './ConfigMapCreate';
 
@@ -24,7 +23,7 @@ export function ConfigMapList(props) {
 
   const description = (
     <Trans i18nKey="config-maps.description">
-      <Link
+      <ExternalLink
         className="bsl-link"
         url="https://kubernetes.io/docs/concepts/configuration/configmap/"
       />
