+++ src/resources/CronJobs/CronJobDescription.js
@@ -1,9 +1,9 @@
 import { Trans } from 'react-i18next';
-import { Link } from '../../shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
 
 export const description = (
   <Trans i18nKey="cron-jobs.description">
-    <Link
+    <ExternalLink
       className="bsl-link"
       url="https://kubernetes.io/docs/concepts/workloads/controllers/cron-jobs/"
     />
