+++ src/resources/CustomResourceDefinitions/CustomResourceDefinitionList.js
@@ -1,8 +1,7 @@
-import React from 'react';
 import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
 import { Tokens } from 'shared/components/Tokens';
 
 import { CustomResourceDefinitionCreate } from './CustomResourceDefinitionCreate';
@@ -26,7 +25,7 @@ export function CustomResourceDefinitionList(props) {
 
   const description = (
     <Trans i18nKey="custom-resource-definitions.description">
-      <Link
+      <ExternalLink
         className="bsl-link"
         url="https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/"
       />
