+++ src/resources/StorageClasses/StorageClassList.js
@@ -1,15 +1,14 @@
-import React from 'react';
 import { Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
 
 import { StorageClassCreate } from './StorageClassCreate';
 
 export function StorageClassList(props) {
   const description = (
     <Trans i18nKey="storage-classes.description">
-      <Link
+      <ExternalLink
         className="bsl-link"
         url="https://kubernetes.io/docs/concepts/storage/storage-classes/"
       />
