+++ src/shared/components/EmptyListComponent/EmptyListComponent.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 import { useTranslation, Trans } from 'react-i18next';
 import { Button, IllustratedMessage } from '@ui5/webcomponents-react';
 import '@ui5/webcomponents-fiori/dist/illustrations/AllIllustrations';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
 import './EmptyListComponent.scss';
 import { spacing } from '@ui5/webcomponents-react-base';
 
@@ -58,7 +58,7 @@ export const EmptyListComponent = ({
             </Button>
           )}
           {url && (
-            <Link
+            <ExternalLink
               className="emptyListComponent__link bsl-link"
               text="Learn More"
               url={url}
